refactor tests directory to initial structure
