+++ server/src/tests/user.test.js
@@ -3,7 +3,7 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 
-import server from '../../index';
+import server from '../index';
 
 chai.should();
